+++ FuckFantasy12/Assets/Scripts/TestingHansi/Player/PlayerCharController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 92987e8d64bb75946a3130ac97cb6732
+timeCreated: 1495279055
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
