+++ FuckFantasy12/Assets/Scripts/TestingHansi/Player/PlayerInput.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: d4093d9bf29348f419db69aa71a043d2
-timeCreated: 1494963339
+guid: d255ef000edc8d24d8e413f34703ce9c
+timeCreated: 1495271326
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
