+++ FuckFantasy12/FuckFantasy12.csproj
@@ -78,8 +78,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Assets\Scripts\TestingHansi\Player\CharController.cs" />
     <Compile Include="Assets\Scripts\TestingHansi\Player\MouseController.cs" />
+    <Compile Include="Assets\Scripts\TestingHansi\Player\PlayerCharController.cs" />
+    <Compile Include="Assets\Scripts\TestingHansi\Player\PlayerInput.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="GenerateTargetFrameworkMonikerAttribute" />
